Moved UDPMessage and PacketBuilder classes into separate files. Added documentation for UDPMessage, UDPServer and PacketBuilder classes.
